fix(HTD-35): move data to section
